Some work to get films in a grid #26 - but no good
Becasuse our max width is only 710px I tried a new page layout
but that was quite messy to fix up
So tryng a flex grid on a width that is only 710 wide means only
two columns is any use, and that's not all that good.
All our images are different sizes so look bad in a grid.

Just a bigger piece of work